+++ src/components/common/InputsUserInfo/InputProfileImage/InputProfileImage.tsx
@@ -7,7 +7,7 @@ import { useDropzone } from 'react-dropzone';
 import defaultProfileImage from '../../../../assets/images/defaultProfile.png';
 import { Theme } from '../../../../styles/Theme';
 import Button from '../../Button/Button';
-import Label from '../../../design/Label/Label';
+import Label from '../../Label/Label';
 import ProfileImage from '../../Profile/ProfileImage/ProfileImage';
 import {
   profileImageSelectContainer,
